+++ CHANGES.md
@@ -7,6 +7,7 @@ Version 0.5.0
 To be released.
 
  -  Added `-f`/`--stop-on-fail` option.
+ -  Added `-p`/`--parallel` option.
 
 
 Version 0.4.0